refactored component to new file
