+++ apps/app/src/pages/Home/events.ts
@@ -5,8 +5,8 @@ export const events: EventWithSpeakerDto[] = [
     id: 1,
     name: 'Mastering React',
     description: 'An advanced dive into React concepts.',
-    startsAt: '2025-01-16T08:00:00Z',
-    endsAt: '2025-01-16T23:00:00Z',
+    startsAt: '2025-01-19T08:00:00Z',
+    endsAt: '2025-01-19T23:00:00Z',
     maxParticipants: 100,
     requirements: 'Basic React knowledge',
     footageLink: undefined,
@@ -70,8 +70,8 @@ export const events: EventWithSpeakerDto[] = [
     id: 1,
     name: 'Mastering React',
     description: 'An advanced dive into React concepts.',
-    startsAt: '2025-01-16T08:00:00Z',
-    endsAt: '2025-01-16T23:00:00Z',
+    startsAt: '2025-01-19T08:00:00Z',
+    endsAt: '2025-01-19T23:00:00Z',
     maxParticipants: 100,
     requirements: 'Basic React knowledge',
     footageLink: undefined,
