+++ apps/app/src/pages/Home/sections/EventsSection/index.ts
@@ -1,0 +1,3 @@
+import EventsSection from './EventsSection';
+
+export default EventsSection;
